Properly convert long message length -- fixes #11 -- bump version to 1.2.1

+++ src/main/java/tech/gusavila92/websocketclient/common/Utils.java
@@ -20,10 +20,12 @@ public class Utils {
 	/**
 	 * Converts the int value passed as parameter to a 8 byte array
 	 * 
+     * NB: java int is 32 bits, need to take long as an argument (it is 64 bits)
+     *
 	 * @param value
 	 * @return
 	 */
-	public static byte[] to8ByteArray(int value) {
+	public static byte[] to8ByteArray(long value) {
 		return new byte[] { (byte) (value >>> 56), (byte) (value >>> 48), (byte) (value >>> 40), (byte) (value >>> 32),
 				(byte) (value >>> 24), (byte) (value >>> 16), (byte) (value >>> 8), (byte) value };
 	}
